WINTERMUTE: Move wme-plugins implementations to /ext/ folder
