+++ engines/wintermute/base/base_game.cpp
@@ -54,6 +54,7 @@
 #include "engines/wintermute/base/scriptables/script_stack.h"
 #include "engines/wintermute/base/scriptables/script.h"
 #include "engines/wintermute/base/sound/base_sound.h"
+#include "engines/wintermute/ext/plugins.h"
 #include "engines/wintermute/video/video_player.h"
 #include "engines/wintermute/video/video_theora_player.h"
 #include "engines/wintermute/utils/utils.h"
@@ -3147,26 +3148,6 @@ bool BaseGame::externalCall(ScScript *script, ScStack *stack, ScStack *thisStack
 	}
 
 	//////////////////////////////////////////////////////////////////////////
-	// SteamAPI
-	//////////////////////////////////////////////////////////////////////////
-	else if (strcmp(name, "SteamAPI") == 0) {
-		thisObj = thisStack->getTop();
-
-		thisObj->setNative(makeSXSteamAPI(_gameRef,  stack));
-		stack->pushNULL();
-	}
-
-	//////////////////////////////////////////////////////////////////////////
-	// WMEGalaxyAPI
-	//////////////////////////////////////////////////////////////////////////
-	else if (strcmp(name, "WMEGalaxyAPI") == 0) {
-		thisObj = thisStack->getTop();
-
-		thisObj->setNative(makeSXWMEGalaxyAPI(_gameRef,  stack));
-		stack->pushNULL();
-	}
-
-	//////////////////////////////////////////////////////////////////////////
 	// Object
 	//////////////////////////////////////////////////////////////////////////
 	else if (strcmp(name, "Object") == 0) {
@@ -3410,6 +3391,12 @@ bool BaseGame::externalCall(ScScript *script, ScStack *stack, ScStack *thisStack
 		stack->pushNative(arr, false);
 
 		delete[] copy;
+	}
+#endif
+
+#if 1
+	else if(!DID_FAIL(EmulatePluginCall(_gameRef, stack, thisStack, name))) {
+		return STATUS_OK;
 	}
 #endif
