+++ engines/wintermute/ext/wme_galaxy.h
