+++ engines/wintermute/ext/wme_steam.h
