+++ engines/wintermute/persistent.cpp
@@ -81,8 +81,8 @@
 #include "engines/wintermute/base/scriptables/script_ext_mem_buffer.h"
 #include "engines/wintermute/base/scriptables/script_ext_object.h"
 #include "engines/wintermute/base/scriptables/script_ext_string.h"
-#include "engines/wintermute/base/scriptables/script_ext_steam_api.h"
-#include "engines/wintermute/base/scriptables/script_ext_wme_galaxy_api.h"
+#include "engines/wintermute/ext/wme_steam.h"
+#include "engines/wintermute/ext/wme_galaxy.h"
 #include "engines/wintermute/ui/ui_button.h"
 #include "engines/wintermute/ui/ui_edit.h"
 #include "engines/wintermute/ui/ui_entity.h"
@@ -158,6 +158,7 @@ void SystemClassRegistry::registerClasses() {
 	REGISTER_CLASS(SXMemBuffer, false)
 	REGISTER_CLASS(SXObject, false)
 	REGISTER_CLASS(SXString, false)
+
 	REGISTER_CLASS(SXSteamAPI, false)
 	REGISTER_CLASS(SXWMEGalaxyAPI, false)
